были учтены замечания, но планируется в будущем выполнить новый коммит после того, как я пересмотрю получившейся вариант
