--- sementsov aleksandr/BioinformStrong-LEXF/data/rosalind_lexf.txt
@@ -1,2 +1,0 @@
-A B C
-12
