--- sementsov aleksandr/BioinformStrong-LEXF/src/rosalind_lexf.txt
@@ -1,2 +1,0 @@
-A B C
-12
