+++ sementsov aleksandr/bioionformatika/.idea/description.html
@@ -1,0 +1,1 @@
+<html>Simple <b>Java</b> application that includes a class with <code>main()</code> method</html>
