+++ sementsov aleksandr/bioionformatika/.idea/project-template.xml
@@ -1,0 +1,3 @@
+<template>
+  <input-field default="com.company">IJ_BASE_PACKAGE</input-field>
+</template>
