+++ sementsov aleksandr/bioionformatika/rosalind-DNA.txt
@@ -1,0 +1,1 @@
+AGCTTTTCATTCTGACTGCAACGGGCAATATGTCTCTGTGTGGATTAAAAAAAGAGTGTCTGATAGCAGC
