+++ sementsov aleksandr/bioionformatika/rosalind-HAMM.txt
@@ -1,0 +1,2 @@
+GAGCCTACTAACGGGAT
+CATCGTAATGACGGCCT
