+++ sementsov aleksandr/bioionformatika/rosalind-SUBS.txt
