+++ sementsov aleksandr/bioionformatika/rosalind_LEXF.txt
@@ -1,0 +1,2 @@
+A B C
+9
