+++ sementsov aleksandr/bioionformatika/src/BioinformStrong/SUBS/output.txt
